refactor: move summariseTroliboxes and getContainerColors funcs into seperate lib files
