+++ src/components/content/Container/ContainerSVG.tsx
@@ -1,6 +1,7 @@
 import * as Sentry from '@sentry/browser';
 import { Suspense, lazy } from 'preact/compat';
 
+import getContainerColours from '@/lib/getContainerColours';
 import { ContainerName } from '@/types/locatorApi';
 
 // The blank svg takes up the same space as a wheeled bin would whilst the icon is loading
@@ -30,55 +31,6 @@ const containerNameToSvgName: { [key in ContainerName]: string } = {
   'Trolibox – 3 containers': 'trolibox-multi',
 };
 
-const boxPositionMap = {
-  'Trolibox - Top box': 'top',
-  'Trolibox - Middle box': 'middle',
-  'Trolibox - Bottom box': 'bottom',
-};
-
-const getPositionColours = (
-  position: string,
-  lidColor: string,
-  bodyColor: string,
-) => ({
-  [`--lid-${position}-colour`]: lidColor,
-  [`--body-${position}-colour`]: bodyColor,
-});
-
-const getContainerColours = (
-  name: ContainerName,
-  lidColour: string,
-  bodyColour: string,
-  colors?: { [key: string]: { [key: string]: string } },
-) => {
-  const positions = Object.keys(boxPositionMap);
-  if (name !== 'Trolibox – 3 containers' && !positions.includes(name)) {
-    return {
-      '--lid-colour': lidColour ?? 'transparent',
-      '--body-colour': bodyColour,
-    };
-  }
-
-  if (name === 'Trolibox – 3 containers') {
-    return Object.assign(
-      {},
-      ...positions.map((position) =>
-        getPositionColours(
-          boxPositionMap[position],
-          colors[position]['lidColour'] ?? 'transparent',
-          colors[position]['bodyColour'],
-        ),
-      ),
-    );
-  }
-
-  return getPositionColours(
-    boxPositionMap[name],
-    lidColour ?? 'transparent',
-    bodyColour,
-  );
-};
-
 export default function ContainerSvg({
   name,
   lidColour,
@@ -101,7 +53,7 @@ export default function ContainerSvg({
     );
   });
 
-  const cssVariables = getContainerColours(name, lidColour, bodyColour, colors);
+  const cssVariables = getContainerColours(name, bodyColour, lidColour, colors);
 
   return (
     <Suspense fallback={BlankSvg}>
